backend/extraction_service/requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
sqlalchemy>=2.0.43
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0
pytesseract>=0.3.13
pdf2image>=1.17.0
opencv-python>=4.12.0.88
Pillow>=10.0.0
tika>=2.6.0
redis>=6.4.0